+++ contracts/interfaces/PowerIndexPoolInterface.sol
@@ -30,5 +30,5 @@ interface PowerIndexPoolInterface is BPoolInterface {
       uint256 targetDenorm
     );
 
-  function getMinWeight() external view override returns (uint);
+  function getMinWeight() external view override returns (uint256);
 }
